Rework relalg.xcore EMF model to have the containment hierarchy represented by the metamodel definition for operators and expressions.
For NamedElement's (variables, labels etc.) it is still needed to add explicitly to the container
as their instances are usually used multiple times by operators and expressions in the model instance graph.

Thgough operator and expression instances might eventually have multiple parents, this is not a usual sitoation,
and will not cause problems.