cypher2relalg: Allow consistent handling of dontCare variables in the compiler.
dontCare variables are the ones that has no name given in the cypher query.
The cypher2relalg compiler generates a name for them. This patch ensures that if used more than once,
they receive the same name.

This is useful when a variable needs to be visited by the compiler multiple times.
E.g. when compiling the pattern expression below, the rightmost node variable and the edge variable
needs to be visited twice: first to create an outer join, second to create an IS_NOT_NULL selection operator
that refer to the same variable.
WHERE (n1)-[:Foo]->(:Bar)