cypher2relalg: all Variable's in the relalg model are now comparable.
Only AttributeVariable was comparable due to the structure of the relational algebra model.
Now all variables (i.e. also VertexVariable and EdgeVariable instances) are comparable expressions.
This allows for equality test of two variables.

However, it does not make sense to compare e.g. an edge variable to a numeric expression.
The cypher to relational algebra compiler should emit a warning or error. See: #37.