cypher2relalg: introduce expression containers for variables.
VariableExpression is a generic container for a variable to make it easy
to include into expressions.

Various sub-types like VariableComparableExpression, VariableArithmeticExpression
exists to allow type-safe building of expressions. However keep in mind, that
it is not guaranteed that the variable will hold a value that is e.g.
comparable in case of a VariableComparableExpression.
Note: type check is always needed in query execution time!