cypher2relalg: Variable no longer is a subclass of Atom, so is not an Expression anymore.
The relational algebra EMF model has recently been reworked to allow
type-safe packaging of variables and function invocations
present in expression locations, e.g. a variable in a comparable expression position
is wrapped into a VariableComparableExpression, a function invocation in
an arithmetic expression position is wrapped into a FunctionArithmeticExpression.

To continue this, variables in general expression position are
also wrapped into the general VariableExpression.

This change made it unnecessary for a Variable to be descendant
of Expression (through Atom), so this commit also removes this.