cypher2relalg: re-work MATCH ... WHERE subclause handling.
Selection compiled from the WHERE filter of a MATCH clause should be put above
the join that connects several MATCH clauses.

For OPTIONAL MATCH, filter is added to the left outer join itself
as it is solely part of the join criteria.
To handle this, ThetaLeftOuterJoinOperator is introduced to the relalg model.