Fix inconsistency in constraint satisfaction
Previously it was possible for a bound constraint to be bound a second time. This was because implied constraints have
not been considered during binding, and whether an implied constraint was free/bound at the time of satisfying the original one was not determined.

This has been changed to the following:
- Initially every constraint is free.
- For an operation to be applicable every direct or indirect constraint comprising the post-condition criteria that is not directly or indirectly bound by the pre-condition criteria MUST be free.
- When a constraint is bound, all of its implications are bound too, thus they will be no longer free.

Also, a bug in the constraint comparator resulted in undefined behaviour when two constraints differed only in parameters.