Refactor bind-op to support matching future and present operations incrementally
Previously bind-op matched operations for both free and bound constraints. The resulting list consists of operations
that are immediately applicable only (called present operations). We can however do a partial match on free
constraints only, which results in a list of non-past operations, a superset of present operations. In a later step
matching for bound constraints can be incrementally applied.
(Also feasible to match for non-future operations first, but the doesn't make sense partically.)

Code is getting complicated sadly :(