Prune columns from cells having the same constraint adornment
Previously it was possible to have cells in each column that have the same constraint adornment. This means the
algorithm retains multiple search trajectories leading to the same subpattern. This is something that is unfavourable because
 - they are unnecessary: they lead to the same PM search subspace, there is no way that a worse subplan will lead to a better plan in the future than its rival
 - they eat up memory and calculation: they fill the table with unnecessary entries
 - they force out other subpattern candidates from the table detrimenting optimization

Deduplicating columns was implemented with having a lookup table by constraint adornment in each column. On a
prospective insertion, the lookup table is checked and if a duplicate is found the one with the less cost is retained.