Refactor constraint lookup
Yeyy! The ubiquitous all-pervasive "ConstraintLookup" is getting a name and getting out of compiler. Additionally to
segrating it which will largely enhance readability of the search plan calculation it gets an efficient lookup-by-var
functionality. ie you can look up constraints by the variables they contain. How cool is that?

(Probably it will be useful for estimation.)