Flexible variable bindings in nested tasks
The architecture of the task executor was clumsy at best. Tasks got their variable names from the OpBindings. Not just
that it doesn't even make sense to assign tasks to OpBindings but Ops, Also because the names were fixed there was no
straightforward way to remap the variable names in nested scopes. Such nested scope is created inside a pattern (call),
and since the actual name to which the inner name have to be rebound (the name of the variable to be read / written)
is only known when the pattern is called, this made the code tangled in a total mess.

This commit removes some idiocity:

- Task and Weight are assigned to Op not OpBinding

- ISearch altered to have [this bindings variables ctx] as parameters

Previously 'variables' and any context-like stuff was under state. This was bad because we know we have to propagate
variables from parent to child in the search tree, however it is not clear how to propagate any other context:
Should the whole tree get a single context? Should it be passed from parents the children like variables? Should
they be passed from each search-step to another in a strict sequential manner killing all parallelization opportunities?
We don't know yet because we don't know what is in 'ctx' anyways. Currently it is only the Indexer, and as long as it
is NOT MUTATED BY ANY SEARCH STEP the whole tree can get a singleton of it. Possibly later we might need all 3 different
contexts differing in the way they are passed on to successive steps, who knows? So long, variables are getting the fuck
out of 'context' and it becomes a singleton. This means search steps return only variables, and no context.

Also notice 'bindings' Yepp, variable names are now bound pretty late: when the task is called.

- Renamings regarding the compiler and Java interface removed

This removes the Scala test as well. The compiler is no longer exposed, as patterns will become the main abstraction.
There will be a Java API for compiling patterns, but that will be in a different file, under a different name.

I still don't know how to implement unification rules.