Rework CSV loader to use custom classes
The CSV loader load data to custom CsvVertex/CsvEdge (in lieu of the Neo4j driver's InternalNode/InternalRelationship classes).
This allows the driver to attach metadata to the source and target vertex ids, namely, their labels.
This way, vertex ids are not required to be globally unique, instead, it is sufficient if they are unique for a given vertex label.