Cypher2GPlan: Re-work clause compilation for query parts. Fixes #347
This changed query parts compilation to handle query parts purely
as a sequence of clauses.

Previously, clause semantics were taken into account w/ some heuristics
during compilation. It worked as long as query part splitting was in line with
the heuristics.

This change to the compiler makes it unrelated to the way
a query is split to query parts.